refactor(metadata): the sidecar pull primitives are their own leaf
cover-service imported them from series-backfill; once the backfill's
stale-cover refresh became a request to the cover service that edge was a
cycle, and a partially-mocked backfill inside it resolved to the wrong
instance under test (the real pullMokuroEntry ran in the publish suite).
Neither module imports the other any more.
